updated 10-add.c

+++ 0x02-functions_nested_loops/10-add.c
@@ -11,7 +11,7 @@
 
 int add(int a, int b)
 {
-	long r;
+	unsigned int r;
 
 	r = a + b;
 	return (r);
